Add documentation, examples, don't use reqwest
Reqwest is an issue since it shares a header type with Hyper, but
this library supports a version of Hyper newer than the one used
by Reqwest.

I may end up making separate packages for Hyper, Rocket, and Reqwest
support, but it seems silly to have 4 packages for what ammounts
to a few lines of code.